Add ability to handle flag 'mayfail' in database migrations.
Some database migrations depend on the content of the tables and these
could be changed by the user, for example document types.

In order to not break a migration because an update script fails, the
flag # mayfail was added. If it is set before an SQL statement, this
specific statement may fail and that does not break the update.

Just prepend the statement that must not succeed with a line
INSERT INTO...