Abort Akonadi start if it is already broken
If Akonadi is already marked broken Akonadi::Control::start()
will not return breaking the initialization of Kraft.
Add special handling until Akonadi is fixed.

On ArchLinux this happens right now because kraft is using an older
version of libakonadi that works with qt5 while the akonadi daemon needs
to use a newer one with qt6.

See:
* https://invent.kde.org/pim/akonadi/-/merge_requests/189
* https://aur.archlinux.org/packages/kraft#comment-976037