Extract Git info at build time only if possible
The build system tries to extract some Git build details (e.g. branch
and SHA). The problem is that it assumes that both 'git' is installed,
and that the sources are in a Git repository: while this is true during
a Git build, they are both not true when building from release tarballs
(e.g. as done by distributions).

Hence, manually search for 'git' and that the current sources are a Git
repository:
- if both are valid, extract the build details as done so far
- if neither is available, still try to use '.tag' for the SHA, setting
  the value for the branch as well to it